requirements.txt
adium-theme-ubuntu==0.3.4
attrs==16.3.0
chardet==2.3.0
click==6.6
colorama==0.3.7
constantly==15.1.0
cryptography==1.7.1
deluge==1.3.15
dj-database-url==0.4.2
Django==1.11.6
enum34==1.1.6
gnome-gmail==2.3.1
gunicorn==19.7.1
idna==2.2
incremental==16.10.1
ipaddress==1.0.17
numpy==1.12.1
PAM==0.4.2
Pillow==4.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.7
pygame===1.9.1release
pygobject==3.22.0
pygpgme==0.3
pyOpenSSL==16.2.0
pyserial==3.2.1
python-libtorrent==1.1.1
pytz==2017.2
pyxdg==0.25
scikit-learn==0.19.0
scipy==0.19.1
service-identity==16.0.0
setproctitle==1.1.10
six==1.10.0
sklearn==0.0
Twisted==16.6.0
unity-lens-photos==1.0
whitenoise==3.3.1
zope.interface==4.3.2
